Implement support for 32-bit ASNs

+++ pkg/agent/apis/types.go
@@ -218,7 +218,7 @@ func (r ServiceExternalIPInfo) SortRows() bool {
 type BGPPolicyResponse struct {
 	BGPPolicyName string `json:"name,omitempty"`
 	RouterID      string `json:"routerID,omitempty"`
-	LocalASN      int32  `json:"localASN,omitempty"`
+	LocalASN      int64  `json:"localASN,omitempty"`
 	ListenPort    int32  `json:"listenPort,omitempty"`
 }
 
@@ -237,7 +237,7 @@ func (r BGPPolicyResponse) SortRows() bool {
 // BGPPeerResponse describes the response struct of bgppeers command.
 type BGPPeerResponse struct {
 	Peer  string `json:"peer,omitempty"`
-	ASN   int32  `json:"asn,omitempty"`
+	ASN   int64  `json:"asn,omitempty"`
 	State string `json:"state,omitempty"`
 }
